+++ TurboBudget/Core/Network/Components/AsyncButton.swift
@@ -1,44 +1,44 @@
+////
+////  AsyncButton.swift
+////  NetworkTemplate
+////
+////  Created by Theo Sementa on 03/10/2024.
+////
 //
-//  AsyncButton.swift
-//  NetworkTemplate
+//import SwiftUI
 //
-//  Created by Theo Sementa on 03/10/2024.
+//struct AsyncButton<Label: View>: View {
+//    
+//    // Builder
+//    let action: () async -> Void
+//    let label: () -> Label
+//    
+//    @State private var isLoading = false
+//    
+//    // init
+//    init(action: @escaping () async -> Void, @ViewBuilder label: @escaping () -> Label) {
+//        self.action = action
+//        self.label = label
+//    }
+//    
+//    // MARK: -
+//    var body: some View {
+//        Button {
+//            Task {
+//                isLoading = true
+//                await action()
+//                isLoading = false
+//            }
+//        } label: {
+//            label()
+//        }
+//        .disabled(isLoading)
+//    } // End body
+//} // End struct
 //
-
-import SwiftUI
-
-struct AsyncButton<Label: View>: View {
-    
-    // Builder
-    let action: () async -> Void
-    let label: () -> Label
-    
-    @State private var isLoading = false
-    
-    // init
-    init(action: @escaping () async -> Void, @ViewBuilder label: @escaping () -> Label) {
-        self.action = action
-        self.label = label
-    }
-    
-    // MARK: -
-    var body: some View {
-        Button {
-            Task {
-                isLoading = true
-                await action()
-                isLoading = false
-            }
-        } label: {
-            label()
-        }
-        .disabled(isLoading)
-    } // End body
-} // End struct
-
-// MARK: - Preview
-#Preview {
-    AsyncButton {} label: {
-        Text("COUCOU")
-    }
-}
+//// MARK: - Preview
+//#Preview {
+//    AsyncButton {} label: {
+//        Text("COUCOU")
+//    }
+//}
