+++ TurboBudget/Core/Network/Generic/NetworkService+Extensions.swift
@@ -1,19 +1,19 @@
+////
+////  NetworkService.swift
+////  LifeFlow
+////
+////  Created by Theo Sementa on 09/03/2024.
+////
 //
-//  NetworkService.swift
-//  LifeFlow
+//import NetworkKit
 //
-//  Created by Theo Sementa on 09/03/2024.
-//
-
-import NetworkKit
-
-extension NetworkService {
-    
-    @MainActor
-    static func handleError(error: Error) {
-        if let error = error as? NetworkError {
-            BannerManager.shared.banner = error.banner
-        }
-    }
-    
-}
+//extension NetworkService {
+//    
+//    @MainActor
+//    static func handleError(error: Error) {
+//        if let error = error as? NetworkError {
+//            BannerManager.shared.banner = error.banner
+//        }
+//    }
+//    
+//}
