+++ TurboBudget/Core/Network/Managers/BannerManager.swift
@@ -1,14 +1,14 @@
+////
+////  BannerManager.swift
+////  Essential
+////
+////  Created by KaayZenn on 11/03/2024.
+////
 //
-//  BannerManager.swift
-//  Essential
+//import Foundation
 //
-//  Created by KaayZenn on 11/03/2024.
-//
-
-import Foundation
-
-class BannerManager: ObservableObject {
-    static let shared = BannerManager()
-    
-    @Published var banner: Banner?
-}
+//class BannerManager: ObservableObject {
+//    static let shared = BannerManager()
+//    
+//    @Published var banner: Banner?
+//}
