+++ TurboBudget/Epic/Accounts/Network/AccountAPIRequester.swift
@@ -7,6 +7,7 @@
 
 import Foundation
 import NetworkKit
+import CoreModule
 
 enum AccountAPIRequester: APIRequestBuilder {
     case fetch
