+++ TurboBudget/Epic/Authentification/Screens/LoginScreen.swift
@@ -44,17 +44,17 @@ struct LoginScreen: View {
             
             #if DEBUG
             Button {
-                Task {
-                    do {
-                        let user = try await NetworkService.sendRequest(apiBuilder: UserAPIRequester.login(
-                            email: "test.user@example.com",
-                            password: "password"
-                        ), responseModel: UserModel.self)
+//                Task {
+//                    do {
+//                        let user = try await NetworkService.sendRequest(apiBuilder: UserAPIRequester.login(
+//                            email: "test.user@example.com",
+//                            password: "password"
+//                        ), responseModel: UserModel.self)
 //                        TokenManager.shared.setTokenAndRefreshToken(token: user.token ?? "", refreshToken: user.refreshToken ?? "")
-                        UserStore.shared.currentUser = user
-                        AppManager.shared.appState = .success
-                    } catch { }
-                }
+//                        UserStore.shared.currentUser = user
+//                        AppManager.shared.appState = .success
+//                    } catch { }
+//                }
             } label: {
                 Text("Login mocks")
             }
