+++ TurboBudget/Epic/Subscriptions/Network/SubscriptionAPIRequester.swift
@@ -7,6 +7,7 @@
 
 import Foundation
 import NetworkKit
+import CoreModule
 
 enum SubscriptionAPIRequester: APIRequestBuilder {
     case fetch(accountID: Int)
