+++ TurboBudget/Epic/User/Network/UserService.swift
@@ -1,20 +1,20 @@
+////
+////  UserService.swift
+////  CashFlow
+////
+////  Created by Theo Sementa on 29/04/2025.
+////
 //
-//  UserService.swift
-//  CashFlow
+//import Foundation
+//import NetworkKit
 //
-//  Created by Theo Sementa on 29/04/2025.
-//
-
-import Foundation
-import NetworkKit
-
-struct UserService {
-    
-    static func update(body: UserModel) async throws -> UserModel {
-        return try await NetworkService.sendRequest(
-            apiBuilder: UserAPIRequester.update(body: body),
-            responseModel: UserModel.self
-        )
-    }
-    
-}
+//struct UserService {
+//    
+//    static func update(body: UserModel) async throws -> UserModel {
+//        return try await NetworkService.sendRequest(
+//            apiBuilder: UserAPIRequester.update(body: body),
+//            responseModel: UserModel.self
+//        )
+//    }
+//    
+//}
