+++ UserModule/Sources/Models/AuthBody.swift
@@ -1,0 +1,13 @@
+//
+//  AuthBody.swift
+//  UserModule
+//
+//  Created by Theo Sementa on 24/07/2025.
+//
+
+
+import Foundation
+
+struct AuthBody: Codable {
+    var identityToken: String
+}
